clarified file hierarchy in README (#39)
Had a bunch of issues trying to combine threshold keys that I generated.
Documentation was quite unclear so I had to go through the examples
(still couldn't figure it out). Reached out to @jazkamer on discord
(@selfuryon?) and told me that my wallet directoriies needed to have the
same name as the peers specified in the configs. Hopefully this
clarifies for future users. This is a really useful tool.